+++ new_lineage_panel/src/ActionWidget.tsx
@@ -1,16 +1,6 @@
-import {
-  Dispatch,
-  FC,
-  FunctionComponent,
-  PropsWithChildren,
-  SetStateAction,
-  useContext,
-  useEffect,
-  useState,
-} from "react";
+import { PropsWithChildren, useContext, useEffect, useState } from "react";
 import "reactflow/dist/style.css";
-import { Button, Card, CardBody, Input, Label, Tooltip } from "reactstrap";
-import AlertCircleIcon from "./assets/icons/alert-circle.svg?react";
+import { Button, Card, CardBody } from "reactstrap";
 import ResetIcon from "./assets/icons/reset.svg?react";
 import HelpIcon from "./assets/icons/help.svg?react";
 import FeedbackIcon from "./assets/icons/feedback.svg?react";
@@ -34,26 +24,6 @@ import {
 import classNames from "classnames";
 import { BetterPopover } from "./components/Modal";
 import { DEFAULT_MIN_ZOOM } from "./utils";
-
-const InfoIcon: FunctionComponent<{ id: string; message: string }> = ({
-  id,
-  message,
-}) => {
-  const [isOpen, setIsOpen] = useState(false);
-  return (
-    <div
-      className={styles.alert_icon}
-      id={id}
-      onMouseEnter={() => setIsOpen(true)}
-      onMouseLeave={() => setIsOpen(false)}
-    >
-      <AlertCircleIcon />
-      <Tooltip target={id} isOpen={isOpen}>
-        {message}
-      </Tooltip>
-    </div>
-  );
-};
 
 const ActionButton = ({
   onClick,
@@ -260,15 +230,8 @@ const AutoExpansionPopover = () => {
   );
 };
 
-export const ActionWidget: FC<{
-  selectCheck: boolean;
-  setSelectCheck: Dispatch<SetStateAction<boolean>>;
-  nonSelectCheck: boolean;
-  setNonSelectCheck: Dispatch<SetStateAction<boolean>>;
-}> = ({ selectCheck, setSelectCheck, nonSelectCheck, setNonSelectCheck }) => {
+export const ActionWidget = () => {
   const {
-    selectedColumn,
-    confidence,
     setSidebarScreen,
     setSelectedColumn,
     setCollectColumns,
@@ -282,74 +245,6 @@ export const ActionWidget: FC<{
         <CardBody className={styles.menu_card}>
           <div className="d-flex gap-sm">
             <AutoExpansionPopover />
-            {aiEnabled && selectedColumn.name && (
-              <>
-                <div className={styles.select_node_checkbox}>
-                  <Input
-                    type="checkbox"
-                    id="select-check"
-                    className="mt-0"
-                    checked={selectCheck}
-                    onChange={(e) => {
-                      if (CLL.inProgress) {
-                        CLL.showCllInProgressMsg();
-                        return;
-                      }
-                      setSelectCheck(e.target.checked);
-                    }}
-                  />
-                  <Label check for="select-check">
-                    Select
-                  </Label>
-                  <InfoIcon
-                    id="select_lineage"
-                    message="Select linkages are shown if there is direct flow of data between columns through select statements."
-                  />
-                </div>
-                <div className={styles.non_select_node_checkbox}>
-                  <Input
-                    type="checkbox"
-                    id="non-select-check"
-                    className="mt-0"
-                    checked={nonSelectCheck}
-                    onChange={(e) => {
-                      if (CLL.inProgress) {
-                        CLL.showCllInProgressMsg();
-                        return;
-                      }
-                      setNonSelectCheck(e.target.checked);
-                    }}
-                  />
-                  <Label check for="non-select-check">
-                    Non-Select
-                  </Label>
-                  <InfoIcon
-                    id="non_select_lineage"
-                    message={
-                      "Non-Select linkages are shown if columns appear " +
-                      "in condition/clauses like where, join, having, etc."
-                    }
-                  />
-                </div>
-                {confidence.confidence === "low" && (
-                  <>
-                    <div className={styles.verticle_divider} />
-                    <div className="d-flex gap-xxs align-items-center">
-                      <div>Confidence</div>
-                      <InfoIcon
-                        id="confidence"
-                        message={
-                          "Depending on the SQL dialect and complexity of queries, " +
-                          "there may be situations where we are not completely " +
-                          "confident about the lineage shown in this view"
-                        }
-                      />
-                      <div className={styles.low_confidence}>Low</div>
-                    </div>
-                  </>
-                )}
-              </>
-            )}
           </div>
         </CardBody>
       </Card>
